Refactor : 미사용 로직 삭제
