+++ src/main/java/com/catchroom/chat/feign/dto/AccommodationResponse.java
@@ -1,5 +1,6 @@
-package com.catchroom.chat.message.dto;
+package com.catchroom.chat.feign.dto;
 
+import com.catchroom.chat.message.dto.RoomResponse;
 import com.catchroom.chat.message.type.AccommodationType;
 import java.util.List;
 import lombok.AllArgsConstructor;
